Simulate a dataset compliant with AMS-I.B

Generates a tidy dataset representing mechanical energy services delivered by renewable systems that displace fossil fuel-based equipment.

simulate_ams_ib_dataset(
  n_machines = 15,
  n_periods = 12,
  start_year = 2023,
  start_month = 1,
  mean_service_mj = 48000,
  sd_service_mj = 6000,
  emission_factor = 0.00007,
  project_fossil_share = 0.05
)

Arguments

n_machines

Number of mechanical systems to simulate.

n_periods

Number of monitoring periods per system.

start_year

Calendar year of the first monitoring period.

start_month

Calendar month (1-12) of the first monitoring period.

mean_service_mj

Mean annual mechanical energy service per machine in MJ.

sd_service_mj

Standard deviation of annual mechanical energy service in MJ.

emission_factor

Baseline emission factor in tCO2e/MJ.

project_fossil_share

Share of mechanical service retained by fossil back-up systems (0-1).

Value

A tibble containing monitoring metadata, baseline fuel characteristics, and emission outcomes.

Examples

simulate_ams_ib_dataset(n_machines = 5)
#> # A tibble: 60 × 16
#>    machine_id monitoring_period  year month   day monitoring_date
#>    <chr>                  <int> <dbl> <dbl> <int> <date>         
#>  1 machine_1                  1  2023     1    21 2023-01-21     
#>  2 machine_1                  2  2023     2    12 2023-02-12     
#>  3 machine_1                  3  2023     3     2 2023-03-02     
#>  4 machine_1                  4  2023     4     6 2023-04-06     
#>  5 machine_1                  5  2023     5     4 2023-05-04     
#>  6 machine_1                  6  2023     6    16 2023-06-16     
#>  7 machine_1                  7  2023     7     2 2023-07-02     
#>  8 machine_1                  8  2023     8     3 2023-08-03     
#>  9 machine_1                  9  2023     9    25 2023-09-25     
#> 10 machine_1                 10  2023    10    22 2023-10-22     
#> # ℹ 50 more rows
#> # ℹ 10 more variables: monitoring_label <chr>, service_output_mj <dbl>,
#> #   fuel_consumption <dbl>, net_calorific_value <dbl>, emission_factor <dbl>,
#> #   project_energy_mj <dbl>, baseline_energy_mj <dbl>,
#> #   baseline_emissions_tco2e <dbl>, project_emissions_tco2e <dbl>,
#> #   emission_reductions_tco2e <dbl>